api-gateway/requirements.txt
# SaaS Factory API Gateway - Standalone Requirements
# Night 56: API Gateway dependencies for end-to-end testing

# Web Framework & API
fastapi==0.115.14
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.0
starlette==0.42.0

# HTTP Client & WebSocket
httpx==0.28.1
websockets>=13.0,<15.1.0
requests==2.32.3

# Database & ORM
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.2.0

# Google Cloud Platform (basic)
google-cloud-logging==3.12.1
google-cloud-secret-manager==2.24.0
google-auth==2.40.3
google-cloud-bigquery

# AI & ML
openai==1.93.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==45.0.5

# Utilities
python-json-logger==2.0.7
pyyaml==6.0.2
python-multipart==0.0.17
email-validator==2.2.0
jinja2==3.1.6
aiofiles==24.1.0

# Background Tasks
celery==5.4.0

# External Services
stripe==12.3.0
sendgrid==6.11.0

# Configuration
python-decouple==3.8
